server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aeventkit==2.1.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.15",
    "aiomysql==0.2.0",
    "aiosignal==1.4.0",
    "aiosmtplib==4.0.2",
    "alembic==1.16.4",
    "alpaca-py==0.42.1",
    "androguard>=4.1.3",
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "asyncpg==0.30.0",
    "asyncssh>=2.22.0",
    "attrs==25.3.0",
    "bcrypt==4.3.0",
    "beautifulsoup4==4.13.5",
    "boto3>=1.42.84",
    "cachetools==5.5.2",
    "certifi==2025.8.3",
    "cffi==2.0.0",
    "chardet==5.2.0",
    "charset-normalizer==3.4.3",
    "click==8.2.1",
    "cryptography>=46.0.5",
    "cssselect==1.3.0",
    "cssutils==2.11.1",
    "curl-cffi==0.13.0",
    "dnspython==2.7.0",
    "email-validator==2.2.0",
    "emails==0.6",
    "fastapi==0.116.1",
    "fastapi-cli==0.0.8",
    "fastapi-cloud-cli==0.1.5",
    "frozendict==2.4.6",
    "frozenlist==1.7.0",
    "google-api-core==2.25.1",
    "google-api-python-client==2.179.0",
    "google-auth==2.40.3",
    "google-auth-httplib2==0.2.0",
    "google-auth-oauthlib==1.2.2",
    "googleapis-common-protos==1.70.0",
    "greenlet==3.2.4",
    "gunicorn==23.0.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httplib2==0.22.0",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "ib-async==2.0.1",
    "idna==3.10",
    "jinja2==3.1.6",
    "kuzu==0.11.2",
    "ldap3>=2.9.1",
    "lxml==6.0.1",
    "mako==1.3.10",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "more-itertools==10.7.0",
    "msgpack==1.1.1",
    "multidict==6.6.4",
    "multitasking==0.0.12",
    "neo4j==5.28.2",
    "nest-asyncio==1.6.0",
    "numpy==2.3.3",
    "oauthlib==3.3.1",
    "packaging==25.0",
    "pandas==2.3.2",
    "peewee==3.18.2",
    "platformdirs==4.4.0",
    "premailer==3.10.0",
    "propcache==0.3.2",
    "proto-plus==1.26.1",
    "protobuf==6.32.0",
    "psycopg2-binary==2.9.10",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.23",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.10.1",
    "pygments==2.19.2",
    "pyjwt==2.10.1",
    "pymysql==1.1.2",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.20",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "redis>=7.2.1",
    "reportlab>=4.4.10",
    "requests==2.32.5",
    "requests-oauthlib==2.0.0",
    "rich==14.1.0",
    "rich-toolkit==0.15.0",
    "rignore==0.6.4",
    "rsa==4.9.1",
    "sentry-sdk==2.35.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.8",
    "sqlalchemy==2.0.43",
    "sseclient-py==1.8.0",
    "starlette==0.47.3",
    "stripe==12.4.0",
    "surrealdb==1.0.8",
    "ta==0.11.0",
    "typer==0.16.1",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "uritemplate==4.2.0",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
    "uvloop==0.21.0",
    "watchfiles==1.1.0",
    "websockets==15.0.1",
    "yarl==1.20.1",
    "yfinance==0.2.66",
]
